requirements.txt
# requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
python-decouple
whitenoise
dj-database-url